:recycle: refactor: melhorar reset de paginação

+++ semana_3/src/app/shared/utils/abstract-list/abstract-list.ts
@@ -1,13 +1,11 @@
-import { inject } from "@angular/core";
-
-import { TableLazyLoadEvent } from "primeng/table";
+import { Table, TableLazyLoadEvent } from "primeng/table";
+import { ConfirmationService, MessageService } from "primeng/api";
 
 import { PaginationProperties } from "../../../interfaces/pagination.interfaces";
-import { ConfirmationService } from "primeng/api";
 
-export abstract class AbstractList<T> {
-  private readonly _confirmationService: ConfirmationService =
-    inject(ConfirmationService);
+export abstract class AbstractList<T extends Object> {
+  protected abstract _messageService: MessageService;
+  protected abstract _confirmationService: ConfirmationService;
 
   public abstract readonly columns: {
     key: keyof T;
@@ -21,11 +19,7 @@ export abstract class AbstractList<T> {
   public searchTerm = "";
   public paginationProperties: PaginationProperties = {};
 
-  constructor() {}
-
-  public abstract fetchItens(): void;
-  public abstract onEditItem(item: T): void;
-  public abstract onDeleteItem(item: T): void;
+  public abstract tableRef?: Table;
 
   private updatePageAndSize({ first, rows }: TableLazyLoadEvent): void {
     const page = first ?? this.currentPage;
@@ -37,16 +31,43 @@ export abstract class AbstractList<T> {
     this.fetchItens();
   }
 
-  public openDeleteDialog(item: T): void {
-    this._confirmationService.confirm({
-      message: "Deseja realmente deletar o item de id: " + (item as any)["id"],
-      accept: () => this.onDeleteItem(item),
+  private showErrorMessage(detail: string) {
+    this._messageService.add({
+      severity: "error",
+      summary: "Erro!",
+      detail,
     });
   }
 
-  public search(): void {
+  private resetTablePagination(): void {
     this.currentPage = 1;
 
+    if (this.tableRef) {
+      this.tableRef.first = 1;
+    }
+  }
+
+  public abstract fetchItens(): void;
+  public abstract onEditItem(item: T): void;
+  public abstract onDeleteItem(item: T): void;
+
+  public openDeleteDialog(item: T): void {
+    if ("id" in item) {
+      this._confirmationService.confirm({
+        message: "Deseja realmente deletar o item de id: " + item["id"],
+        accept: () => this.onDeleteItem(item),
+      });
+
+      return;
+    }
+
+    this.showErrorMessage(
+      "Ops! Houve um problema ao tentar deletar este item!"
+    );
+  }
+
+  public search(): void {
+    this.resetTablePagination();
     this.fetchItens();
   }
